MainWindows refactoring + along side bug fix
