+++ workers.py
@@ -1,11 +1,22 @@
 from PyQt6.QtCore import QThread, pyqtSignal
+from PyQt6.QtWidgets import QTableWidgetItem
 
 import time
 import registry
 import hardware
 import system
+import data
+import support
 
+import localization
+import Entities
+
+trans = localization.trans
+languages = Entities.Languages()
+
+# 
 # Воркер для обновления информации в трее
+# 
 class TrayWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -29,7 +40,11 @@ class TrayWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+
+
+# 
 # Воркер для сбора данных
+# 
 class CollectSlowDataWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -54,7 +69,23 @@ class CollectSlowDataWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+#Создаем воркер для сбора и обновления информации
+def startCollectSlowDataWorker(self):
+    self.collect_slow_worker = CollectSlowDataWorker(self)
+    self.collect_slow_worker.result.connect(self.processSlowData)
+    self.collect_slow_worker.start()
+
+#Обработка данных из collectWorker'а
+def processSlowData(self, result):        
+    data.writeTempData(self, result)
+    data.writeTDPData(self, result)
+    data.writeCoresData(self, result)
+
+
+
+# 
 # Воркер для мониторинга нагрузки на CPU
+# 
 class CollectFastDataWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -82,7 +113,27 @@ class CollectFastDataWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+#Создаем воркер для мониторинга нагрузки на CPU
+def startCollectFastDataWorker(self):
+    self.collect_fast_worker = CollectFastDataWorker(self)
+    self.collect_fast_worker.result.connect(self.processFastData)
+    self.collect_fast_worker.start()
+
+#Обработка данных из CollectFastDataWorker'а
+def processFastData(self, result):
+    if not self.config.getIsCPUManagmentOn():
+        CPU_performance_mode = False
+    else:
+        CPU_performance_mode = hardware.setCpuPerformanceState(self.config, self.data_lists)
+
+    self.config.setPerformanceCPUModeOn(CPU_performance_mode)
+
+    data.writeThreadsData(self, result)
+
+
+# 
 # Воркер для бэкапа данных
+# 
 class BackupWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -111,7 +162,21 @@ class BackupWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+#Создаем воркер для бэкапа данных
+def startBackupWorker(self):
+    self.backup_worker = BackupWorker(self)
+    self.backup_worker.result.connect(self.saveStatistics)
+    self.backup_worker.start()
+
+#Обработка данных из backupWorker'а
+def saveStatistics(self):
+    support.saveStatistics(self)
+
+
+
+# 
 # Воркер для мониторинга системных параметров
+# 
 class SystemMonitoringWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -137,7 +202,26 @@ class SystemMonitoringWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+#Создаем воркер для мониторинга системных параметров
+def startSystemMonitoringWorker(self):
+    self.system_monitoring_worker = SystemMonitoringWorker(self)
+    self.system_monitoring_worker.result.connect(self.updateSystemStateConfig)
+    self.system_monitoring_worker.start()
+
+#Обработка данных из SystemMonitoringWorker'а
+def updateSystemStateConfig(self, data):
+    self.config.setSystemUsesLightTheme(data['system_uses_light_theme'])
+
+    support.writeToConfig(self.config)
+    new_config = support.readConfig(self)
+
+    self.collect_slow_worker.update(new_config)
+
+
+
+# 
 # Воркер для обновления показателей в интерфейсе
+# 
 class UpdateUiScoresWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -159,7 +243,69 @@ class UpdateUiScoresWorker(QThread):
     def stop(self):
         self.keepRunning = False
 
+def startUpdateUiScoresWorker(self):
+    #Создаем воркер обновления показателей интерфейса
+    #У нас теперь данные стекаются с разных воркеров, так что обновлять надо централизовано
+    self.update_ui_scores_worker = UpdateUiScoresWorker(self)
+    self.update_ui_scores_worker.result.connect(self.updateUiScores)
+    self.update_ui_scores_worker.start()
+
+#Обработка данных из UpdateUiScoresWorker'а — обновляем показатели в интерфейсе
+def updateUiScores(self, result):
+
+    data_lists = self.data_lists
+    
+    if len(data_lists['general_temps']) > 0:
+        #Минимальная
+        self.lineEditCpuMinTemp.setText(str(data_lists['min_temp']))
+        #Текущая
+        self.lineEditCpuCurrentTemp.setText(str(data_lists['current_temp']))
+        #Максимальная
+        self.lineEditCpuMaxTemp.setText(str(data_lists['max_temp']))
+
+    if len(data_lists['general_TDP']) > 0:
+        #Минимальный
+        self.lineEditCpuMinTDP.setText(str(data_lists['min_TDP']))
+        #Текущий
+        self.lineEditCpuCurrentTDP.setText(str(data_lists['current_TDP']))
+        #Максимальный
+        self.lineEditCpuMaxTDP.setText(str(data_lists['max_TDP']))
+
+    if len(data_lists['average_temps']) > 0:
+
+        row = 0
+
+        for minutes in (1, 5, 15, 60, 60*24):
+            avg_temp = support.toRoundStr(hardware.getAvgTempForSeconds(self, 60*minutes))
+            avg_tdp = support.toRoundStr(hardware.getAvgTDPForSeconds(self, 60*minutes))
+
+            self.tableAverage.setItem(row, 0, QTableWidgetItem(f"{ avg_temp } С°"))            
+            self.tableAverage.setItem(row, 1, QTableWidgetItem(f"{ avg_tdp } { trans(self.locale, 'watt') }"))
+
+            row += 1
+
+    for core in self.data_lists['cpu']['cores']:
+        self.CPUinfoTable.setItem(core['id'], 0, QTableWidgetItem(core['clock']))
+
+    #Идем по физическим ядрам, потому что нам надо сопоставить нагрузку между потоками и ядрами   
+    for index, thread in enumerate(data_lists['cpu']['threads']):
+        if self.SMT:
+            load = support.toRoundStr((
+                                        float(data_lists['cpu']['threads'][(thread['id']-1)*2]['load']) + 
+                                        float(data_lists['cpu']['threads'][(thread['id']-1)*2-1]['load'])
+                                        )/2)
+        else:
+            load = str(thread['load'])
+
+        self.CPUinfoTable.setItem(thread['id'], 1, QTableWidgetItem(f"{ load }%"))
+
+        if self.SMT and index + 1 == (self.cpu_threads/2):
+            break
+
+
+# 
 # Воркер для обновления изображения в трее
+# 
 class UpdateTrayIconWorker(QThread):
     def __init__(self, app_self):
         super().__init__()
@@ -180,3 +326,54 @@ class UpdateTrayIconWorker(QThread):
 
     def stop(self):
         self.keepRunning = False
+
+#Создаем воркер обновления иконки в трее
+def startUpdateTrayIconWorker(self):
+    self.update_tray_icon_worker = UpdateTrayIconWorker(self)
+    self.update_tray_icon_worker.result.connect(self.updateTrayIcon)
+    self.update_tray_icon_worker.start()
+
+#Обработка данных из UpdateTrayIconWorker'а
+def updateTrayIcon(self, result):
+    data_lists = self.data_lists
+
+    if len(data_lists['general_temps']) > 0:
+        #Сформируем новое изображения трея
+        self.image = support.getTrayImage(data_lists['current_temp'], self.config)
+
+def updateWorkersCollectionInterval(self):
+    self.collect_slow_worker.update(self.config)
+    self.update_ui_scores_worker.update(self.config)
+    self.update_tray_icon_worker.update(self.config)
+    self.backup_worker.update(self.config)
+
+def startWorkers(self):
+    #Запустим воркеры для сбора данных
+    startCollectSlowDataWorker(self)
+    startCollectFastDataWorker(self)
+
+    #Для обновления иконки трея
+    startUpdateTrayIconWorker(self)
+
+    #Для обновления GUI
+    startUpdateUiScoresWorker(self)
+
+    #Для бэкапа
+    if self.config.getOpenMinimized:
+        self.update_ui_scores_worker.stop()
+
+    if self.config.getIsBackupNeeded():
+        if support.getRestoredData(self):
+            self.data_lists = support.getRestoredData(self)
+
+        startBackupWorker(self)
+
+    #И для мониторинга системных данных
+    startSystemMonitoringWorker(self)
+
+def stopWorkers(self):
+    self.collect_fast_worker.stop()
+    self.collect_slow_worker.stop()
+    self.update_ui_scores_worker.stop()
+    self.backup_worker.stop()
+    self.system_monitoring_worker()
